+++ delivery-info-service/src/main/resources/application.yml
@@ -1,12 +1,5 @@
 server:
   port: 8888
 
-spring:
-  profiles:
-    active: production
-
-docker:
-  image:
-    version: -1
 hello:
-  world: Hello World! - delivery-info-service/${docker.image.version}
+  world: Hello World! - delivery-info-service from here
