+++ delivery-relay-service/Dockerfile
@@ -9,10 +9,10 @@ ARG VERSION
 # docker run --e 로 오버라이딩
 ENV COMMAND_LINE_ARGS_BEFORE=${COMMAND_LINE_ARGS_BEFORE}
 ENV COMMAND_LINE_ARGS_AFTER=${COMMAND_LINE_ARGS_AFTER}
-ENV SERVER_PORT=8888
+ENV SERVER_PORT=8090
 ENV VERSION=${VERSION}
 
 ADD ${PACKAGE_JAR_LOCATION} app.jar
 
-EXPOSE ${SERVER_PORT}:8888
+EXPOSE ${SERVER_PORT}:8090
 ENTRYPOINT java ${COMMAND_LINE_ARGS_BEFORE} ${VERSION} -jar ./app.jar ${COMMAND_LINE_ARGS_AFTER}
