+++ delivery-relay-service/src/main/resources/application-production.yml
@@ -1,10 +1,6 @@
 server:
   port: 8090
 
-restClient:
-  proxy:
-    baseUrl: /relay/v1
-
 spring:
   autoconfigure:
     exclude: org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration,org.springframework.boot.autoconfigure.data.mongo.MongoDataAutoConfiguration,org.springframework.boot.autoconfigure.mongo.embedded.EmbeddedMongoAutoConfiguration
