+++ delivery-relay-service/src/main/resources/application-test.yml
@@ -1,7 +1,2 @@
 server:
   port: 8090
-
-restClient:
-  proxy:
-    baseUrl: /relay/v1
-
