+++ delivery-relay-service/src/main/resources/application.yml
@@ -1,6 +1,5 @@
-docker:
-  image:
-    version: -1
+server:
+  port: 8090
 
 hello:
-  world: Hello World! from Delivery-Relay-Service v.${docker.image.version}
+  world: Hello World! from Delivery-Relay-Service .from Here
