+++ readme.md
@@ -8,22 +8,18 @@
 
 <br>
 
-### :bicyclist: 하루 100만 주문을 해결하는 배달의민족은 어떻게 만들어진 것일까요?
+### :bicyclist: 프로젝트 목표
 
-- 일반적인 클라이언트-서버의 웹서버를 넘어 여러 주체(사장님,고객님 등..)와 어떻게 통신을 주고받을까요?
-- 여러 서비스들의 사이에서 어떻게 장애를 다룰까요?
-- 트래픽 최저점시간 대비 100배까지 증가하는 점심, 저녁 피크타임 트래픽 스파이크를 어떻게 해결할까요?
-
-이러한 궁금증을 해결하기 위해서 배달의민족 중계서버를 구현하게 되었습니다.
-
-<br>
-
-### :bicyclist: 무엇을 경험하고 싶은건가요?
-
-- 예상치 못한 장애들을 핸들링하는 경험하기
-- 여러 각도에서 테스트를 진행하고 테스트 코드를 작성하며 버그 최소화하는 경험하기
-- 모든 기능이 탑재된 서비스를 개발하는 것이 아닌, 진화하는 서비스를 경험하기
-- 분산환경에서 여러 서비스들과 통신하는 과정을 경험하기
+- [ ] 여러 각도에서 테스트를 진행하고 테스트 코드를 작성하며 버그 최소화하는 경험하기 (= 테스트 CI&CD)
+    - [x] 단위테스트 instruction & branch 커버리지 100%
+    - [x] 단위테스트 커버리지 체크 자동화
+    - [ ] 다양한 어플리케이션 레벨 테스트 작성 (ex. Load Test, Stress Test, 여러 유저 시나리오 등...)
+    - [ ] 어플리케이션 레벨 테스트 자동화 (w/ AWS)
+- [ ] 예상 못한 장애들을 핸들링하는 경험하기 (= Disaster Engineering)
+    - [ ] Chaos Engineering 툴 적용하기
+- [ ] 모든 기능이 탑재된 서비스를 개발하는 것이 아닌, 진화하는 서비스를 경험하기 (= 프로젝트 회고)
+- [ ] 분산환경에서 여러 서비스들과 통신하는 과정을 경험하기 (= Microservices && interconnection)
+- [ ] 트래픽 최저점시간 대비 100배까지 증가하는 점심, 저녁 피크타임 트래픽 스파이크를 어떻게 해결할까요? (= Scaling && 테스트 시나리오)
 
 <br>
